+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Removed direct dependency on JUnit 4, so you can choose more freely which JUnit version to use.
 * Added `ifTake` and `unlessTake` inversions of `takeIf` and `takeUnless`.
 * Added `runIf` and `applyIf`.
+* Added `testTimeout: Duration = 1.minutes` param to `runTest()`.
 
 ## 5.13.0